requirements.txt
flask>=3.0.0
pyTigerGraph>=1.3.0
python-dotenv>=1.0.0
